Split platform-specific code into separate per-platform files
Extract Windows, Linux, and stub implementations from monolithic files
into dedicated platform files, following the existing RHI backend pattern:

- ConsoleProcessManager.cpp (847 lines) split into:
  - ConsoleProcessManager.cpp (shared: constructor, ProcessCommands, CommandRegistry)
  - ConsoleProcessManagerWin32.cpp (Win32 pipes, CreateProcess)
  - ConsoleProcessManagerLinux.cpp (POSIX pipes, fork/exec)
  - ConsoleProcessManagerStub.cpp (unsupported platform fallbacks)

- PlatformInput.cpp (1434 lines) split into:
  - PlatformInput.cpp (shared: key names, PlatformInputManager)
  - PlatformInputWin32.cpp (Win32 raw input, XInput gamepad)
  - PlatformInputSDL2.cpp (SDL2 event polling, SDL_GameController)

- Platform.h (1500 lines) split into:
  - Platform.h (compiler/platform detection, ~210 lines)
  - PlatformTypes.h (Win32 type aliases, macros, HRESULT)
  - PlatformAudioStubs.h (XAudio2, XInput stubs)
  - PlatformDirectXMathStubs.h (XMVECTOR, XMMATRIX, math ops)
  - PlatformD3DStubs.h (D3D11 forward decls, ComPtr, DXGI enums)

No CMake changes needed — GLOB_RECURSE auto-discovers new files,
and each platform file self-guards with #ifdef preprocessor macros.
