+++ wiki/Home.md
@@ -123,12 +123,12 @@ SparkEngine is licensed under the [MIT License](https://github.com/Krilliac/Spar
 <!-- AUTO:stats -->
 | Metric | Count |
 |--------|-------|
-| Header files | 371 |
+| Header files | 375 |
 | ECS Components | 40 |
 | ECS Systems | 46 |
 | Editor Panels | 32 |
 | Test files | 84 |
 | Test cases | 1052+ |
 | Wiki pages | 55 |
-| *Last synced* | *2026-03-18 20:43* |
+| *Last synced* | *2026-03-18 21:32* |
 <!-- /AUTO:stats -->
